refactor: 파일 구조 및 이름 리팩토링(#274)
