--- src/main/java/PodoeMarket/podoemarket/dto/response/OrderPerformanceListDTO.java
@@ -1,18 +1,0 @@
-package PodoeMarket.podoemarket.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class OrderPerformanceListDTO {
-    private LocalDateTime createdAt;
-    List<OrderPerformanceDTO> orderPerformance;
-}
