+++ src/main/java/PodoeMarket/podoemarket/service/VerificationService.java
@@ -11,7 +11,7 @@ import java.time.Duration;
 @Slf4j
 @Service
 @RequiredArgsConstructor
-public class RedisUtil {
+public class VerificationService {
     private final StringRedisTemplate redisTemplate;//Redis에 접근하기 위한 Spring의 Redis 템플릿 클래스
 
     public String getData(String key){//지정된 키(key)에 해당하는 데이터를 Redis에서 가져오는 메서드
@@ -21,11 +21,6 @@ public class RedisUtil {
         } catch(Exception e){
             return "redis get 문제 발생";
         }
-    }
-
-    public void setData(String key,String value){//지정된 키(key)에 값을 저장하는 메서드
-        ValueOperations<String,String> valueOperations=redisTemplate.opsForValue();
-        valueOperations.set(key,value);
     }
 
     public void setDataExpire(String key,String value,long duration){//지정된 키(key)에 값을 저장하고, 지정된 시간(duration) 후에 데이터가 만료되도록 설정하는 메서드
