+++ src/main/java/PodoeMarket/podoemarket/user/controller/UserController.java
@@ -2,14 +2,14 @@ package PodoeMarket.podoemarket.user.controller;
 
 import PodoeMarket.podoemarket.Utils.ValidCheck;
 import PodoeMarket.podoemarket.common.config.jwt.JwtProperties;
-import PodoeMarket.podoemarket.dto.EmailCheckDTO;
-import PodoeMarket.podoemarket.dto.EmailRequestDTO;
+import PodoeMarket.podoemarket.user.dto.request.EmailCheckRequestDTO;
+import PodoeMarket.podoemarket.user.dto.request.EmailRequestDTO;
 import PodoeMarket.podoemarket.dto.response.ResponseDTO;
 import PodoeMarket.podoemarket.dto.UserDTO;
 import PodoeMarket.podoemarket.common.entity.UserEntity;
 import PodoeMarket.podoemarket.common.security.TokenProvider;
-import PodoeMarket.podoemarket.mail.MailSendService;
-import PodoeMarket.podoemarket.service.RedisUtil;
+import PodoeMarket.podoemarket.service.MailSendService;
+import PodoeMarket.podoemarket.service.VerificationService;
 import PodoeMarket.podoemarket.user.dto.request.SignInRequestDTO;
 import PodoeMarket.podoemarket.user.dto.response.SignInResponseDTO;
 import PodoeMarket.podoemarket.user.dto.response.TokenResponseDTO;
@@ -38,7 +38,7 @@ public class UserController {
     private final UserService userService;
     private final TokenProvider tokenProvider;
     private final MailSendService mailService;
-    private final RedisUtil redisUtil;
+    private final VerificationService verificationService;
     private final JwtProperties jwtProperties;
     private final PasswordEncoder pwdEncoder = new BCryptPasswordEncoder();
 
@@ -137,7 +137,7 @@ public class UserController {
     }
 
     @PostMapping("/mailauthCheck")
-    public ResponseEntity<?> authCheck(@RequestBody @Valid EmailCheckDTO dto){
+    public ResponseEntity<?> authCheck(@RequestBody @Valid EmailCheckRequestDTO dto){
         try{
             boolean Checked = mailService.CheckAuthNum(dto.getEmail(),dto.getAuthNum());
 
@@ -181,7 +181,7 @@ public class UserController {
                     .build();
 
             userService.create(user);
-            redisUtil.deleteData(dto.getAuthNum()); // 인증 번호 확인 후, redis 상에서 즉시 삭제
+            verificationService.deleteData(dto.getAuthNum()); // 인증 번호 확인 후, redis 상에서 즉시 삭제
             mailService.joinSignupEmail(user.getEmail());
 
             return ResponseEntity.ok().body(true);
@@ -271,7 +271,7 @@ public class UserController {
                     .data(userInfo)
                     .build();
 
-            redisUtil.deleteData(dto.getAuthNum()); // 인증 번호 확인 후, redis 상에서 즉시 삭제
+            verificationService.deleteData(dto.getAuthNum()); // 인증 번호 확인 후, redis 상에서 즉시 삭제
             return ResponseEntity.ok().body(resDTO);
         } catch(Exception e) {
             ResponseDTO resDTO = ResponseDTO.builder().error(e.getMessage()).build();
@@ -310,7 +310,7 @@ public class UserController {
                     .refreshToken(tokenProvider.createRefreshToken(user))
                     .build();
 
-            redisUtil.deleteData(dto.getAuthNum()); // 인증 번호 확인 후, redis 상에서 즉시 삭제
+            verificationService.deleteData(dto.getAuthNum()); // 인증 번호 확인 후, redis 상에서 즉시 삭제
             return ResponseEntity.ok().body(resUserDTO);
         } catch(Exception e) {
             ResponseDTO resDTO = ResponseDTO.builder().error(e.getMessage()).build();
